# openapi.yaml
openapi: 3.1.0
info:
  title: SparkyPod API
  description: |
    REST API for managing podcast feeds with automatic ad detection and removal.


    This API allows you to:
    - Manage podcast feed subscriptions
    - View and manage processed episodes
    - Configure ad detection settings including AI model selection
    - Monitor system status and trigger cleanup operations
    - Manage cross-episode ad patterns with network and podcast scope
    - Submit corrections to improve ad detection accuracy
  version: 2.16.0
  contact:
    name: SparkyPod
  license:
    name: MIT
servers:
  # Same-origin relative URL. Swagger UI served from /api/v1/docs uses
  # whatever host the browser is on, so "Try it out" always hits the
  # correct server: localhost in dev, the production domain in prod,
  # a Tailscale/Cloudflare URL on a tunnel, etc. No hardcoded host.
  - url: /api/v1
    description: Same-origin (current server)
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: '{baseUrl}/api/v1'
    description: Custom server
    variables:
      baseUrl:
        default: http://localhost:8000
        description: Base URL of the podcast server
...
paths:
...
  /system/model-pricing:
    get:
      tags:
        - System
      summary: Get model pricing rates
      description: Returns all known LLM model pricing rates from the model_pricing table
      operationId: getModelPricing
      parameters:
        - name: source
          in: query
          required: false
          description: Filter pricing entries by source
          schema:
            type: string
            enum: [openrouter_api, pricepertoken, default, legacy]
      responses:
        '200':
          description: Model pricing list
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    description: All known model pricing entries
                    items:
                      type: object
                      properties:
                        matchKey:
                          type: string
                          description: Normalized model key for pricing lookups
                        rawModelId:
                          type: string
                          description: Original model identifier from pricing source
                        displayName:
                          type: string
                          description: Human-readable model name
                        inputCostPerMtok:
                          type: number
                          description: Input cost per million tokens (USD)
                        outputCostPerMtok:
                          type: number
                          description: Output cost per million tokens (USD)
                        source:
                          type: string
                          description: Pricing data source (openrouter_api, pricepertoken, default, legacy)
                        updatedAt:
                          type: string
                          description: Last updated timestamp
  /system/model-pricing/refresh:
    post:
      tags:
        - System
      summary: Force refresh model pricing
      description: Force refresh pricing data from the active provider's pricing source (OpenRouter API or pricepertoken.com)
      operationId: refreshModelPricing
      responses:
        '200':
          description: Pricing refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  modelsUpdated:
                    type: integer
                    description: Number of models with pricing data
        '502':
          description: Pricing fetch failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /docs:
    get:
      tags:
        - System
      summary: Swagger UI
      description: |
        Renders this OpenAPI spec in an interactive Swagger UI. Assets
        are bundled in the image under `/ui/swagger/*` -- no CDN
        dependency, air-gap friendly. Auth-gated; an unauthenticated
        GET returns 401 (the blueprint-level `check_auth` runs first).
      operationId: swaggerUi
      responses:
        '200':
          description: HTML page with Swagger UI
          content:
            text/html:
              schema:
                type: string
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /docs/:
    get:
      tags:
        - System
      summary: Swagger UI (trailing-slash alias)
      description: Same content as `/docs`; exists so browsers don't
        get a Werkzeug redirect on the trailing slash variant.
      operationId: swaggerUiSlash
      responses:
        '200':
          description: HTML page with Swagger UI
          content:
            text/html:
              schema:
                type: string
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /openapi.yaml:
    get:
      tags:
        - System
      summary: OpenAPI spec
      description: |
        Returns this spec file. Auth-gated. Note that the spec
        describes endpoints relative to `/api/v1`; this handler
        serves the spec at `/api/v1/openapi.yaml`.
      operationId: openapiSpec
      responses:
        '200':
          description: OpenAPI YAML document
          content:
            application/yaml:
              schema:
                type: string
            text/yaml:
              schema:
                type: string
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /system/backup:
    get:
      tags:
        - System
      summary: Download database backup
      description: |
        Returns a copy of the SQLite database file for backup purposes. Rate limited to 6/hour.
      operationId: downloadBackup
      responses:
        '200':
          description: Database backup file download
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'Attachment filename, e.g. attachment; filename="sparkypod-backup-2026-03-12.db"'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Flask session cookie issued by ``POST /auth/login``. Every mutating
        endpoint additionally requires a matching ``X-CSRF-Token`` header
        (double-submit). Public endpoints (``/auth/login``, ``/auth/logout``,
        ``/auth/status``, ``/health``, ``/health/live``) declare ``security: []``.
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          nullable: true
...